meshblock is now a new-style multiblock
