+++ peepingtom/datablocks/multiblocks/__init__.py
@@ -1,4 +1,5 @@
 from .dipoleblock import DipoleBlock
+from .meshblock import MeshBlock
 from .orientedpointblock import OrientedPointBlock
 from .particleblock import ParticleBlock
 from .transformblock import TransformBlock
