+++ peepingtom/datablocks/simpleblocks/__init__.py
@@ -1,6 +1,5 @@
 from .imageblock import ImageBlock
 from .lineblock import LineBlock
-from .meshblock import MeshBlock
 from .orientationblock import OrientationBlock
 from .pointblock import PointBlock
 from .propertyblock import PropertyBlock
