+++ peepingtom/depictors/napari/meshdepictor.py
@@ -3,6 +3,7 @@ from .naparidepictor import NapariDepictor
 
 class MeshDepictor(NapariDepictor):
     def depict(self):
-        self._make_surface_layer(self.datablock.vertices[:, ::-1],
-                                 self.datablock.faces,
+        # .values cause napari does not like xarray yet
+        self._make_surface_layer(self.datablock.vertices.data.values,
+                                 self.datablock.faces.data.values,
                                  name=f'{self.name} - mesh')
